# perf(connector): add integration benchmark for nexmark parsing