--- a/src/connector/benches/nexmark_integration.rs
+++ b/src/connector/benches/nexmark_integration.rs
@@ -1,0 +1,136 @@
+// Copyright 2023 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use criterion::{criterion_group, criterion_main, BatchSize, Criterion};
+use futures::{FutureExt, StreamExt, TryStreamExt};
+use itertools::Itertools;
+use risingwave_common::catalog::ColumnId;
+use risingwave_common::types::DataType;
+use risingwave_connector::parser::{
+    ByteStreamSourceParser, EncodingProperties, JsonParser, JsonProperties, ProtocolProperties,
+    SourceParserIntoStreamExt, SpecificParserConfig,
+};
+use risingwave_connector::source::{
+    BoxSourceStream, BoxSourceWithStateStream, SourceColumnDesc, SourceMessage, SourceMeta,
+    StreamChunkWithState,
+};
+use tracing::Level;
+use tracing_subscriber::prelude::*;
+
+fn make_data_stream() -> BoxSourceStream {
+    let mut generator = nexmark::EventGenerator::default()
+        .with_type_filter(nexmark::event::EventType::Bid)
+        .map(|e| match e {
+            nexmark::event::Event::Bid(bid) => bid,
+            _ => unreachable!(),
+        })
+        .enumerate();
+
+    let message_base = SourceMessage {
+        split_id: "default".into(),
+        key: None,
+        payload: None,
+        offset: "".into(),
+        meta: SourceMeta::Empty,
+    };
+
+    let data_stream = futures::stream::repeat_with(move || {
+        let messages = generator
+            .by_ref()
+            .take(1024)
+            .map(|(i, e)| {
+                let payload = serde_json::to_vec(&e).unwrap();
+                SourceMessage {
+                    payload: Some(payload),
+                    offset: i.to_string(),
+                    ..message_base.clone()
+                }
+            })
+            .collect_vec();
+
+        Ok(messages)
+    });
+
+    data_stream.boxed()
+}
+
+fn make_parser() -> impl ByteStreamSourceParser {
+    let columns = [
+        ("auction", DataType::Int64),
+        ("bidder", DataType::Int64),
+        ("price", DataType::Int64),
+        ("channel", DataType::Varchar),
+        ("url", DataType::Varchar),
+        ("date_time", DataType::Timestamp),
+        ("extra", DataType::Varchar),
+    ]
+    .into_iter()
+    .enumerate()
+    .map(|(i, (n, t))| SourceColumnDesc::simple(n, t, ColumnId::new(i as _)))
+    .collect_vec();
+
+    let props = SpecificParserConfig {
+        key_encoding_config: None,
+        encoding_config: EncodingProperties::Json(JsonProperties {
+            use_schema_registry: false,
+        }),
+        protocol_config: ProtocolProperties::Plain,
+    };
+
+    JsonParser::new(props, columns, Default::default()).unwrap()
+}
+
+fn make_stream_iter() -> impl Iterator<Item = StreamChunkWithState> {
+    let mut stream: BoxSourceWithStateStream =
+        make_parser().into_stream(make_data_stream()).boxed();
+
+    std::iter::from_fn(move || {
+        stream
+            .try_next()
+            .now_or_never()
+            .unwrap()
+            .unwrap()
+            .unwrap()
+            .into()
+    })
+}
+
+fn bench(c: &mut Criterion) {
+    let dispatch: tracing::dispatcher::Dispatch = tracing_subscriber::registry()
+        .with(tracing_subscriber::fmt::layer())
+        .with(
+            tracing_subscriber::filter::Targets::new()
+                .with_target("risingwave_connector", Level::INFO),
+        )
+        .into();
+
+    c.bench_function("parse_nexmark_with_tracing", |b| {
+        b.iter_batched(
+            make_stream_iter,
+            |mut iter| tracing::dispatcher::with_default(&dispatch, || iter.next().unwrap()),
+            BatchSize::SmallInput,
+        )
+    });
+
+    c.bench_function("parse_nexmark", |b| {
+        b.iter_batched(
+            make_stream_iter,
+            |mut iter| iter.next().unwrap(),
+            BatchSize::SmallInput,
+        )
+    });
+}
+
+criterion_group!(benches, bench);
+criterion_main!(benches);