Signed-off-by: Bugen Zhao <i@bugenzhao.com>I hereby agree to the terms of the [RisingWave Labs, Inc. Contributor License Agreement](https://gist.github.com/TennyZhuang/f00be7f16996ea48effb049aa7be4d66#file-rw_cla).

## What's changed and what's your intention?

To help us investigate #12959.

The "integration" means that we are not only evaluating the performance of parser itself (which is already covered by #9195), but also include the related procedure like tracing, offset maintaining, chunk building and so on, which is just like how we do parsing in production.

----

<details>
<summary>Something that made me confused (UPDATE: Resolved but still confused)</summary>

https://github.com/risingwavelabs/risingwave/blob/1ca0a8030cf631886e599894d688d1768501b2a2/src/connector/benches/nexmark_integration.rs#L117-L122

</details>

----

Results:

```
parse_nexmark           time:   [1.2333 ms 1.2371 ms 1.2409 ms]

parse_nexmark_with_tracing
                        time:   [2.0007 ms 2.0038 ms 2.0071 ms]
```

...which confirms https://github.com/risingwavelabs/risingwave/issues/12959#issuecomment-1780750295.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Documentation

- [ ] My PR needs documentation updates. (Please use the **Release note** section below to summarize the impact on users)

## Release note

If this PR includes changes that directly affect users or other significant modifications relevant to the community, kindly draft a release note to provide a concise summary of these changes. Please prioritize highlighting the impact these changes will have on users.


<!--
Please create a release note for your changes.

Discuss technical details in the "What's changed" section, and
focus on the impact on users in the release note.

You should also mention the environment or conditions where the impact may occur.
-->

</details>